feat(charts): wire SimpleCharts CalendarHeatmap into SimpleView
feat(charts): remove DetailedCharts CalendarHeatmap from DetailedView
