+++ app/src/components/Charts/SimpleCharts/CalendarHeatmap/CalendarHeatmap.sql
@@ -1,0 +1,6 @@
+select
+    day::varchar as day,
+    stream_count::double as stream_count
+from ${table}
+where ${year_condition}
+order by day
